Finalize: merge develop into master

+++ src/main/java/com/itwillbs/domain/BOMDetailVO.java
@@ -9,10 +9,10 @@ public class BOMDetailVO {
     private String bomDetailId;
     private String bomId;
     private String materialId;
-    private int qty;
+    private double qty;
     private String unit;
     private String remark;
-    private String materialType; // "육수", "원료"
+    private String materialType; // "육수", "원료" ,"포장재" 
     private Date regDate; //등록일
     
     private String materialName;
